pyproject.toml
[tool.poetry]
name = "auth0-token"
version = "0.0.0" # published version is managed using Git tags (see below)
description = "Retrieves Auth0 JWT tokens using the Authorization Code flow"
authors = ["Kenneth Pronovici <pronovic@ieee.org>"]
license = "Apache-2.0"
readme = "PyPI.md"
homepage = "https://pypi.org/project/auth0-token/"
repository = "https://github.com/pronovic/auth0-token"
include = [
   { path = 'Changelog', format = 'sdist' },
   { path = 'NOTICE', format = 'sdist' },
   { path = 'LICENSE', format = 'sdist' },
   { path = 'README.md', format = 'sdist' },
   { path = 'docs', format = 'sdist' },
   { path = 'tests', format = 'sdist' },
]
packages = [ 
   { include = "auth0token", from = "src" },
]
classifiers=[
   "Programming Language :: Python :: 3",
   "License :: OSI Approved :: Apache Software License",
   "Development Status :: 4 - Beta",
]

# Published version is managed using Git tags
# We get either the tag (like "0.24.1") or a snapshot-type version (like "0.24.1+3.e8319c4")
# If the plugin is not installed, then the version is always "0.0.0", taken from above
[tool.poetry-dynamic-versioning]
enable = true
pattern = '^[vV](?P<base>\d+\.\d+\.\d+)'  # this extracts the version from our vX.Y.Z tag format
format-jinja = "{% if distance == 0 and not dirty %}{{ base }}{% else %}{{ base }}+{{ distance }}.{{ commit }}{% endif %}"

[tool.poetry.scripts]
auth0token = "auth0token.cli:auth0token"

[tool.poetry.dependencies]
python = ">=3.12,<4"
fastapi = "~0, >=0.103.2"
uvicorn = "~0, >=0.23.2"
python-dotenv = "^1.0.1"
asyncio = "^3.4.3"
click = "^8.1.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-testdox = "^3.1.0"
coverage = "^6.5.0"  # coveralls needs v6
pylint = "^3.0.1"
pre-commit = "^3.4.0"
black = "^23.9.1"
mypy = "^1.6.0"
isort = "^5.12.0"
coveralls = "^3.3.1"
colorama = "~0, >=0.4.6"
types-requests = "^2.32.0.20240914"

[tool.black]
line-length = 132
target-version = ['py312']
include = '(src\/scripts\/.*$|\.pyi?$)'
exclude = '''
/(
    \.git
  | __pycache__
  | \.tox
  | \.venv
  | \.poetry
  | build
  | dist
  | docs
  | notes
)/
'''

[tool.isort]
profile = "black"
line_length = 132
skip_glob = [ "docs", "notes" ]

[build-system]
requires = ["poetry>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"
